Revise server after data security @clin1
