+++ scripts/pdf_Kafka/img_getpdf.py
@@ -93,8 +93,8 @@ class img_getpdf(img_integrate.img_integrate):
         if 'pilatus' in self.detector:
             sqfqgr_path = write_pdfgetter(self.process_dir, f'{self.file_name_prefix}_sum', pdfgetter)
 
-        elif 'pe1c' in self.detector:
-            if (self.use_flat_field_pe1c) and ('pe1c' in self.detector):
+        elif 'pe1' in self.detector:
+            if (self.use_flat_field_pe1c) and ('pe1' in self.detector):
                 sqfqgr_path = write_pdfgetter(self.process_dir, f'{self.file_name_prefix}_flat', pdfgetter)
             else:
                 sqfqgr_path = write_pdfgetter(self.process_dir, f'{self.file_name_prefix}_sub', pdfgetter)
