+++ scripts/pdf_Kafka/kafka_server.py
